backup_before_restore_20251128_154813/src/vlm_server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.1.0
httpx>=0.25.0
python-multipart>=0.0.6
